refactor: divide GameView into PlayerGameView and InitViewHelper
